Improve gitver experience
Do not use a separate "meta" config for metadata, since that makes it
confusing to build the project for the first time. Just run gitver
directly on build.

Also wrap it in a script that only replaces the version file if it
changed. This avoids forcing a full source rebuild if nothing changed.